Portofolio_10/deployment/requirements.txt
absl-py==2.1.0
altair==5.3.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.0.1
flatbuffers==24.3.25
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.65.1
h5py==3.11.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.12.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
scipy==1.14.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.37.0
tenacity==8.5.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.14.1
